requirements.txt
# requirements.txt

Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
firebase-admin==6.5.0
google-cloud-pubsub==2.21.1
gunicorn==21.2.0
functions-framework==3.*